% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 5 variables:
\describe{
  \item{DDT}{DDT found in the fish}
  \item{LENGTH}{length in cm}
  \item{MILE}{the mile that the fish was found along}
  \item{RIVER}{river or tributary}
  \item{SPECIES}{species of the fish}
  \item{WEIGHT}{weight of the fish, in gms}
}
}
\usage{
ddt
}
\description{
A data set collected by the US forces from the tennessee river and its tributaries
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
