% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{A function to retrieve information about DDT dataset}
\arguments{
\item{df}{the data frame to be subsetted, which is ddt for this functions purposes}

\item{SPECIES}{the string of species to be to be plotted and subsetted}
}
\value{
plot with length vs weight for species and saves the data frame used in the function to the working directory as LvsWforSPECIES.csv where the "SPECIES" string is replaces with the species used. Also prints to the command line a named list containing the DDT data before subsetting, the subsetted data frame, and a relative frequency table of the river before subsetting.
}
\description{
Returns a plot with length vs weight for species and saves the data frame used in the function to the working directory as LvsWforSPECIES.csv where the "SPECIES" string is replaces with the species used. Also prints to the command line a named list containing the DDT data before subsetting, the subsetted data frame, and a relative frequency table of the river before subsetting.
}
\details{
A function for Project 1 in MATH 4753
}
\examples{
\dontrun{myddt(df=ddt,SPECIES = "CCATFISH")
}
}
